test(atomic): split facet cypress tests into three file
https://coveord.atlassian.net/browse/KIT-3578
